made ports in vdml_adi.c from int32_t to uint8_t

+++ src/devices/vdml_adi.c
@@ -14,63 +14,63 @@
 #include "vdml/registry.h"
 #include "vdml/vdml.h"
 
-adi_port_config_e_t adi_port_get_config(int32_t port) {
+adi_port_config_e_t adi_port_get_config(uint8_t port) {
 	return ext_adi_port_get_config(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_port_get_value(int32_t port) {
+int32_t adi_port_get_value(uint8_t port) {
 	return ext_adi_port_get_value(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_port_set_config(int32_t port, adi_port_config_e_t type) {
+int32_t adi_port_set_config(uint8_t port, adi_port_config_e_t type) {
 	return ext_adi_port_set_config(INTERNAL_ADI_PORT, port, type);
 }
 
-int32_t adi_port_set_value(int32_t port, int32_t value) {
+int32_t adi_port_set_value(uint8_t port, int32_t value) {
 	return ext_adi_port_set_value(INTERNAL_ADI_PORT, port, value);
 }
 
-int32_t adi_analog_calibrate(int32_t port) {
+int32_t adi_analog_calibrate(uint8_t port) {
 	return ext_adi_analog_calibrate(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_analog_read(int32_t port) {
+int32_t adi_analog_read(uint8_t port) {
 	return ext_adi_analog_read(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_analog_read_calibrated(int32_t port) {
+int32_t adi_analog_read_calibrated(uint8_t port) {
 	return ext_adi_analog_read_calibrated(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_analog_read_calibrated_HR(int32_t port) {
+int32_t adi_analog_read_calibrated_HR(uint8_t port) {
 	return ext_adi_analog_read_calibrated_HR(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_digital_read(int32_t port) {
+int32_t adi_digital_read(uint8_t port) {
 	return ext_adi_digital_read(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_digital_get_new_press(int32_t port) {
+int32_t adi_digital_get_new_press(uint8_t port) {
 	return ext_adi_digital_get_new_press(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_digital_write(int32_t port, const bool value) {
+int32_t adi_digital_write(uint8_t port, const bool value) {
 	return ext_adi_digital_write(INTERNAL_ADI_PORT, port, value);
 }
 
-int32_t adi_pin_mode(int32_t port, uint8_t mode) {
+int32_t adi_pin_mode(uint8_t port, uint8_t mode) {
 	return ext_adi_pin_mode(INTERNAL_ADI_PORT, port, mode);
 }
 
-int32_t adi_motor_set(int32_t port, int8_t speed) {
+int32_t adi_motor_set(uint8_t port, int8_t speed) {
 	return ext_adi_motor_set(INTERNAL_ADI_PORT, port, speed);
 }
 
-int32_t adi_motor_get(int32_t port) {
+int32_t adi_motor_get(uint8_t port) {
 	return ext_adi_motor_get(INTERNAL_ADI_PORT, port);
 }
 
-int32_t adi_motor_stop(int32_t port) {
+int32_t adi_motor_stop(uint8_t port) {
 	return ext_adi_motor_stop(INTERNAL_ADI_PORT,port);
 }
